Update Prospecting_FallLeaves.html
Take a look at the changes and see if they make sense. I also fixed any open tags for you.

In terms of the one block where the image wasn't coming in, I couldn't figure out why so I just rebuilt it. I also got rid of the indent on the bullets. I would recommend making the text run on the left side and flipping the image to show on the right side. This will make sure it still looks good on small screens (such as the iPhone 5) where the background color fills up most of the screen. To do this you should be able to copy the left block and paste it under the right one and it should switch the order.

I added the link for the telephone and changed some of the styling on the buttons. I would also recommend putting the gecko in the leaves underneath the get a quote box because there is a harsh straight cutoff that would be hidden nicely at the bottom of the content.

Do you need the email address and message id info at the very bottom? This isn't super common to use.